+++ examples/02_Complex/complex_example.py
@@ -57,10 +57,10 @@ if __name__ == '__main__':
             label='Q_th',  # Thermal output
             bus='Fernwärme',  # Linked bus
             size=fx.InvestParameters(
-                fix_effects=1000,  # Fixed investment costs
+                effects_of_investment=1000,  # Fixed investment costs
                 fixed_size=50,  # Fixed size
                 optional=False,  # Forced investment
-                specific_effects={Costs.label: 10, PE.label: 2},  # Specific costs
+                effects_of_investment_per_size={Costs.label: 10, PE.label: 2},  # Specific costs
             ),
             load_factor_max=1.0,  # Maximum load factor (50 kW)
             load_factor_min=0.1,  # Minimum load factor (5 kW)
@@ -130,7 +130,7 @@ if __name__ == '__main__':
         charging=fx.Flow('Q_th_load', bus='Fernwärme', size=1e4),
         discharging=fx.Flow('Q_th_unload', bus='Fernwärme', size=1e4),
         capacity_in_flow_hours=fx.InvestParameters(
-            piecewise_effects=segmented_investment_effects,  # Investment effects
+            piecewise_effects_of_investment=segmented_investment_effects,  # Investment effects
             optional=False,  # Forced investment
             minimum_size=0,
             maximum_size=1000,  # Optimizing between 0 and 1000 kWh