+++ examples/05_Two-stage-optimization/two_stage_optimization.py
@@ -48,7 +48,9 @@ if __name__ == '__main__':
             Q_fu=fx.Flow(
                 label='Q_fu',
                 bus='Gas',
-                size=fx.InvestParameters(specific_effects={'costs': 1_000}, minimum_size=10, maximum_size=500),
+                size=fx.InvestParameters(
+                    effects_of_investment_per_size={'costs': 1_000}, minimum_size=10, maximum_size=500
+                ),
                 relative_minimum=0.2,
                 previous_flow_rate=20,
                 on_off_parameters=fx.OnOffParameters(effects_per_switch_on=300),
@@ -66,7 +68,9 @@ if __name__ == '__main__':
             Q_fu=fx.Flow(
                 'Q_fu',
                 bus='Kohle',
-                size=fx.InvestParameters(specific_effects={'costs': 3_000}, minimum_size=10, maximum_size=500),
+                size=fx.InvestParameters(
+                    effects_of_investment_per_size={'costs': 3_000}, minimum_size=10, maximum_size=500
+                ),
                 relative_minimum=0.3,
                 previous_flow_rate=100,
             ),
@@ -74,7 +78,7 @@ if __name__ == '__main__':
         fx.Storage(
             'Speicher',
             capacity_in_flow_hours=fx.InvestParameters(
-                minimum_size=10, maximum_size=1000, specific_effects={'costs': 60}
+                minimum_size=10, maximum_size=1000, effects_of_investment_per_size={'costs': 60}
             ),
             initial_charge_state='lastValueOfSim',
             eta_charge=1,