+++ tests/test_functional.py
@@ -146,7 +146,7 @@ def test_fixed_size(solver_fixture, time_steps_fixture):
             Q_th=fx.Flow(
                 'Q_th',
                 bus='Fernwärme',
-                size=fx.InvestParameters(fixed_size=1000, fix_effects=10, specific_effects=1),
+                size=fx.InvestParameters(fixed_size=1000, effects_of_investment=10, effects_of_investment_per_size=1),
             ),
         )
     )
@@ -187,7 +187,7 @@ def test_optimize_size(solver_fixture, time_steps_fixture):
             Q_th=fx.Flow(
                 'Q_th',
                 bus='Fernwärme',
-                size=fx.InvestParameters(fix_effects=10, specific_effects=1),
+                size=fx.InvestParameters(effects_of_investment=10, effects_of_investment_per_size=1),
             ),
         )
     )
@@ -228,7 +228,7 @@ def test_size_bounds(solver_fixture, time_steps_fixture):
             Q_th=fx.Flow(
                 'Q_th',
                 bus='Fernwärme',
-                size=fx.InvestParameters(minimum_size=40, fix_effects=10, specific_effects=1),
+                size=fx.InvestParameters(minimum_size=40, effects_of_investment=10, effects_of_investment_per_size=1),
             ),
         )
     )
@@ -269,7 +269,9 @@ def test_optional_invest(solver_fixture, time_steps_fixture):
             Q_th=fx.Flow(
                 'Q_th',
                 bus='Fernwärme',
-                size=fx.InvestParameters(optional=True, minimum_size=40, fix_effects=10, specific_effects=1),
+                size=fx.InvestParameters(
+                    optional=True, minimum_size=40, effects_of_investment=10, effects_of_investment_per_size=1
+                ),
             ),
         ),
         fx.linear_converters.Boiler(
@@ -279,7 +281,9 @@ def test_optional_invest(solver_fixture, time_steps_fixture):
             Q_th=fx.Flow(
                 'Q_th',
                 bus='Fernwärme',
-                size=fx.InvestParameters(optional=True, minimum_size=50, fix_effects=10, specific_effects=1),
+                size=fx.InvestParameters(
+                    optional=True, minimum_size=50, effects_of_investment=10, effects_of_investment_per_size=1
+                ),
             ),
         ),
     )